# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  github_com_sabbatD_srest-api_internal_lib_userConfig.Meta:
    properties:
      sortBy:
        type: string
      sortOrder:
        type: string
      totalAmount:
        type: integer
    type: object
  github_com_sabbatD_srest-api_internal_lib_userConfig.MetaResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_userConfig.TableUser'
        type: array
      meta:
        $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_userConfig.Meta'
    type: object
  github_com_sabbatD_srest-api_internal_lib_userConfig.PutUser:
    properties:
      email:
        type: string
      phoneNumber:
        type: string
      username:
        maxLength: 60
        minLength: 1
        type: string
    type: object
...
  github_com_sabbatD_srest-api_internal_lib_userConfig.TableUser:
    properties:
      date:
        type: string
      email:
        type: string
      id:
        type: integer
      isAdmin:
        type: boolean
      isBlocked:
        type: boolean
      phoneNumber:
        type: string
      username:
        type: string
    type: object
...
  internal_http-server_handlers_admin.UpdateRequest:
    properties:
      field:
        type: string
      value: {}
    type: object
...
host: easydev.club
info:
  contact:
    email: s4bb4t@yandex.ru
    name: s4bb4t
  description: This is a RESTful API service for EasyDev. It provides various user
    management functionalities such as user registration, authentication, profile
    updates, and admin operations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: sAPI
  version: v0.3.2
paths:
  /admin/users:
    get:
      description: Fetches a list of users based on optional query parameters such
        as filters and sorting.
      parameters:
      - description: Filter users by username or email
        in: query
        name: search
        type: string
      - description: Sort by 'email', 'username', or 'id'. Default is 'id'.
        in: query
        name: sortBy
        type: string
      - description: 'Sort order: ''asc'', ''desc'', or ''none''. Default is ''asc''.'
        in: query
        name: sortOrder
        type: string
      - description: Filter by block status (true/false)
        in: query
        name: isBlocked
        type: boolean
      - description: Limit the number of users returned (default is 20)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default is 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of users.
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_userConfig.MetaResponse'
        "401":
          description: Unauthorized access. Bearer token missing or invalid.
          schema:
            type: string
        "403":
          description: Insufficient permissions.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - admin
  /admin/users/{id}:
    delete:
      description: Deletes a user by their ID.
      parameters:
      - description: ID of the user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User successfully removed.
          schema:
            type: string
        "400":
          description: Invalid or missing user ID.
          schema:
            type: string
        "401":
          description: Unauthorized access. Bearer token missing or invalid.
          schema:
            type: string
        "403":
          description: Insufficient permissions.
          schema:
            type: string
        "404":
          description: User not found.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Remove user
      tags:
      - admin
    get:
      description: Retrieves a user's profile by their ID.
      parameters:
      - description: ID of the user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of user profile.
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_userConfig.TableUser'
        "400":
          description: Invalid or missing user ID.
          schema:
            type: string
        "401":
          description: Unauthorized access. Bearer token missing or invalid.
          schema:
            type: string
        "403":
          description: Insufficient permissions.
          schema:
            type: string
        "404":
          description: User not found.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Retrieve user's profile
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Updates the details of a user by accepting a JSON payload.
      parameters:
      - description: ID of the user
        in: path
        name: id
        required: true
        type: integer
      - description: User data payload
        in: body
        name: UserData
        required: true
        schema:
          $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_userConfig.PutUser'
      produces:
      - application/json
      responses:
        "200":
          description: User profile updated successfully.
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_userConfig.TableUser'
        "400":
          description: Duplicate login or email.
          schema:
            type: string
        "401":
          description: Unauthorized access. Bearer token missing or invalid.
          schema:
            type: string
        "403":
          description: Insufficient permissions.
          schema:
            type: string
        "404":
          description: User not found.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update user's profile
      tags:
      - admin
  /admin/users/{id}/block:
    post:
      description: Blocks a user by their ID, disabling their account.
      parameters:
      - description: ID of the user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User successfully blocked.
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_userConfig.TableUser'
        "400":
          description: Invalid or missing user ID.
          schema:
            type: string
        "404":
          description: User not found.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Block user
      tags:
      - admin
  /admin/users/{id}/rights:
    post:
      consumes:
      - application/json
      description: Updates specific fields related to user's rights by accepting a
        JSON payload.
      parameters:
      - description: ID of the user
        in: path
        name: id
        required: true
        type: integer
      - description: User data for updating rights
        in: body
        name: UserData
        required: true
        schema:
          $ref: '#/definitions/internal_http-server_handlers_admin.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rights successfully updated.
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_userConfig.TableUser'
        "400":
          description: No such field.
          schema:
            type: string
        "404":
          description: User not found.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update user's rights
      tags:
      - admin
  /admin/users/{id}/unlock:
    post:
      description: Unblocks a user by their ID, re-enabling their account.
      parameters:
      - description: ID of the user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User successfully unblocked.
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_userConfig.TableUser'
        "400":
          description: Invalid or missing user ID.
          schema:
            type: string
        "404":
          description: User not found.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Unlock user
      tags:
      - admin
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token required for accessing protected routes. Format:
      "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"